pyproject.toml
[project]
name = "cobot-cli"
version = "0.1.0"
description = "CLI tool to fetch Cobot bookings"
dependencies = [
    "typer",
    "requests",
    "python-dateutil",
    "dynaconf>=3.2.7",
    "python-telegram-bot",
]
requires-python = ">=3.8"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
cobot = "cobot_cli.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cobot_cli"]
